Remove vcftools/0.1.14 and vep/87
They depend on perl/5.22.3, which was compiled on CentOS 6 and has
a broken libmysqlclient.so.16 dependency. There are other programs
that use this perl as well, but I haven't verified whether they a-
lso hit this issue.

Eventually I'd like to remove or re-install all the programs that
were installed using non-system perls.