metaphlan2/2.9.5: Update
Apparently this requires some Python stuff in the environment, they
need mpa_dir to be defined as the install location, and they expect
bowtie2 to be available. Confusingly, the wiki says the variable is
MDIR, but I don't see any references to that in the code.

See: https://bitbucket.org/biobakery/biobakery/wiki/metaphlan2