plasmidid/1.6.3: Add Python stuff
It is not clear in the plasmidID installation instructions that it
needs biopython etc, but the test data fails because pandas is not
installed. I noticed a bunch of python packages in the Bioconda re-
cipe so now I've installed them.

See: https://github.com/bioconda/bioconda-recipes/blob/master/recipes/plasmidid/meta.yaml